requirements.txt
mpu6050-raspberrypi==1.2
pygame==2.6.1
RPi.GPIO==0.7.1
smbus==1.1.post2
smbus2==0.5.0
